+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## [Unreleased]
 
+## [v0.3.0] - 2022-07-31
+
 * Make `spi-flash` `no-std` compatible (still requires `alloc`).
 * The `FlashAccess` trait now has an associated `Error` type and its methods
   return corresponding `Result`s; `From<FlashAccess::Error>` must now be
   implemented for `spi_flash::Error`.
+* Change progress bar style
 
 ## [v0.2.2] - 2021-01-20
 
@@ -29,7 +32,8 @@
 
 * Initial release.
 
-[Unreleased]: https://github.com/adamgreig/spi-flash-rs/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/adamgreig/spi-flash-rs/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/adamgreig/spi-flash-rs/compare/v0.2.2...v0.3.0
 [v0.2.2]: https://github.com/adamgreig/spi-flash-rs/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/adamgreig/spi-flash-rs/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/adamgreig/spi-flash-rs/compare/v0.1.0...v0.2.0